requirements-dev.txt
# Development dependencies with ML support
# Install with: pip install -r requirements-dev.txt

# Base dependencies from requirements.txt
-r requirements.txt

# Parallel testing (pytest-xdist)
pytest-xdist>=3.5.0
pytest-benchmark>=4.0.0

# HTTP client for FastAPI/Starlette TestClient
httpx>=0.24.0

# Machine Learning dependencies
tensorflow>=2.8.0
scikit-learn>=1.0.0
joblib>=1.1.0

# Visualization dependencies
seaborn>=0.11.0
matplotlib>=3.5.0

# Scheduling
schedule>=1.1.0

# Other utilities
pandas>=1.3.0
numpy>=1.21.0

# Database and settings
sqlalchemy>=2.0.0
pydantic-settings>=2.0.0
